update yarn.lock and try to get it working from yarn scripts
