+++ .babelrc.json
@@ -1,3 +1,4 @@
 {
-  "presets": ["@babel/preset-react"]
+  "presets": ["@babel/preset-react"],
+	"plugins": ["@babel/plugin-proposal-optional-chaining"]
 }
